Simlpify, r3f/React only for now
