+++ packages/input-composer/src/useInput.ts
@@ -1,0 +1,6 @@
+import { useGamepad } from "./useGamepad"
+
+export const useInput = () => {
+  const gamepad = useGamepad()
+  return { gamepad }
+}
